Update type from Long to Integer

+++ apex-module/src/main/java/com/charts/apex/entity/ApexObject.java
@@ -11,18 +11,18 @@ public class ApexObject {
 	private String name;
 
 	@JsonProperty("data")
-	private List<Long> values;
+	private List<Integer> values;
 
 	public ApexObject(String name) {
 		this.name = name;
 	}
 
-	public ApexObject(String name, List<Long> values) {
+	public ApexObject(String name, List<Integer> values) {
 		this.name = name;
 		this.values = values;
 	}
 
-	public ApexObject withList(List<Long> values) {
+	public ApexObject withList(List<Integer> values) {
 		this.values = values;
 		return this;
 	}
