+++ recharts-module/src/main/java/com/charts/recharts/entity/RechartsDataObject.java
@@ -6,8 +6,8 @@ import lombok.Data;
 public class RechartsDataObject {
     private String name;
     private String month;
-    private Long value;
-    public RechartsDataObject(String name, String month, Long value) {
+    private Integer value;
+    public RechartsDataObject(String name, String month, Integer value) {
         this.name = name;
         this.month = month;
         this.value = value;
